Add files via upload
For this one, I really got rid of all variable heights, widths, and positioning for the big box containing everything so that it can stay exactly where it needs to. Hopefully, this will make the most sense and fit for all sizes. If this doesn't work, then I'm going to keep the margin-top 8% and margin-left:30-35% as I had it before. We shall see.